REF stats GUI: separate GUI components into files
